test(Calendar): add demo test
