+++ packages/calendar/demo/index.json
@@ -1,0 +1,8 @@
+{
+  "component": true,
+  "usingComponents": {
+    "van-cell": "../../cell/index",
+    "van-calendar": "../../calendar/index",
+    "demo-block": "../../../example/components/demo-block/index"
+  }
+}
